move add whisky to its own file
